+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 *[M]: major change; [m]: minor change*
 
+**Breaking changes**
+
+TBW - separate packages
+
+**Other changes**
+
+* Browser extension:
+    - [m] Add usage hints (popping up when first the app is first used, as well as when the user clicks on the dedicated button)
+
 ## 2.3.1 (December 14, 2016)
 
 * Library: